GCP-Policies/requirements.txt
# Testing
pytest
pytest-cov
pytest-azurepipelines
unittest-xml-reporting

# GCP SDK
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-resource-manager

# Libraries
httplib2